refactor: simplify flow component
