--- src/components/features/registration/ProviderRegistrationFlow/ui/FlowNavigation/index.ts
@@ -1,1 +1,0 @@
-export * from './FlowNavigation';
