+++ src/components/features/registration/ProviderRegistrationFlow/ui/ProviderRegistrationForm/ProviderRegistrationForm.tsx
@@ -1,5 +1,4 @@
 import { Control } from 'react-hook-form';
-
 import { Box } from '@mui/material';
 import { styled } from '@mui/material/styles';
 import { RegisterProvider } from '@/lib/features/registration';
