+++ src/components/features/registration/ProviderRegistrationFlow/ui/index.ts
@@ -1,3 +1,2 @@
-export * from './FlowNavigation';
 export * from './ProviderRegistrationForm';
 export * from './SeatCountForm';
